Initial 'alpha' commit of python server side support
   The compile_typespec script now takes an additional argument
--pyserver=pyserver_filename to specify a filename for the
python back end server. If left unspecified it will default to
ModuleNameServer.py

   The generated python file is currently a combination of the
server and the implementation file. It contains the class
definitions for the implementations as well as a wsgi handler
to take care of dispatching requests.

   The python server file will also run as a standalone wsgi
service and will respond to kbase client requests. It takes
2 arguments: --port and --host
   They are the same as the plackup args with the same name

   Note that the python and perl clients have also been modified
to add an id property to the json-rpc requests. The id is
currently just a random number generated on the fly.