KBASE-922 #comment X-Forwarded-For / X-Real-IP in pl server
Use the X-Forwarded-For and X-Real-IP headers in the perl server as
per KBASE-922:

By default, the logged ip address will be, in order of preference:
1) X-Forwarded-For[0]
2) X-Real-IP
3) The standard address

If the config file for the server includes 
dont_trust_x_ip_headers = true
then only the standard address is logged.

The content of X-Forwarded-For is always logged, if it exists, in a log
message preceding the start method message.